+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- [#93] Delete additional ingress annotations if not present on the dogu resource
+- [#94] Correct ingress annotation in docs and sample
 
 ## [v0.29.0] - 2023-04-06
 ### Added